% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_click_above_for_graph_or_table.R
\name{click_above_for_graph_or_tableUI}
\alias{click_above_for_graph_or_tableUI}
\alias{click_above_for_graph_or_tableServer}
\title{click_above_for_graph_or_table}
\usage{
click_above_for_graph_or_tableUI(id)

click_above_for_graph_or_tableServer(id)
}
\arguments{
\item{id}{id}
}
\description{
click_above_for_graph_or_table
}
\examples{
simpleApp <- function() {
  ui <- fluidPage(
    splshiny::click_above_for_graph_or_tableUI("loader_fig_1")
  )
server <- function(input, output, session) {
  splshiny::click_above_for_graph_or_tableServer("loader_fig_1")
}
shinyApp(ui, server)
}
}
